# swagger.yaml
openapi: 3.0.4
info:
  title: Sigorta Acentesi API
  version: 1.0.0
  description: REST API for managing insurance agency operations
servers:
  - url: https://sigortaapi.elacigdem.com/api/v1
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Notification:
      type: object
      properties:
        id:
          type: integer
          description: Notification ID
        userId:
          type: integer
          description: User ID
        title:
          type: string
          description: Notification title
        message:
          type: string
          description: Notification message
        type:
          type: string
          description: Notification type (email, sms, app)
        status:
          type: string
          description: Notification status (read, unread)
        sentAt:
          type: string
          format: date-time
          description: Date and time when notification was sent
        readAt:
          type: string
          format: date-time
          description: Date and time when notification was read
security:
  - bearerAuth: []
paths:
  /users/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                role:
                  type: string
                  enum: [admin, agent, manager]
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  userId:
                    type: integer
        '400':
          description: Bad request - invalid input
        '500':
          description: Internal server error
  /users/login:
    post:
      summary: Login user and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT token
                  userId:
                    type: integer
                  username:
                    type: string
                  role:
                    type: string
        '401':
          description: Unauthorized - invalid credentials
        '500':
          description: Internal server error
...
  /notifications:
    get:
      summary: List user's notifications
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
        '401':
          description: Unauthorized - missing or invalid token
        '500':
          description: Internal server error
    post:
      summary: Create a new notification
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                title:
                  type: string
                message:
                  type: string
                type:
                  type: string
                  enum: [email, sms, app]
      responses:
        '201':
          description: Notification created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  notificationId:
                    type: integer
        '400':
          description: Bad request - invalid input
        '401':
          description: Unauthorized - missing or invalid token
        '500':
          description: Internal server error
  /notifications/mark-all-read:
    post:
      summary: Mark all notifications as read
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All notifications marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  updatedCount:
                    type: integer
        '401':
          description: Unauthorized - missing or invalid token
        '500':
          description: Internal server error
...